Rename db/interface to db/database
